Restructure runner server folders
